[feat] Add get user daat hook
It can be reused to get user data for user profile page
